新增对钉钉机器人加签方法的支持
