+++ src/lib/ldap.go
@@ -1,23 +1,23 @@
 package lib
 
 import (
-"Yearning-go/src/model"
-"crypto/tls"
-"errors"
-"fmt"
-"gopkg.in/ldap.v3"
+	"Yearning-go/src/model"
+	"crypto/tls"
+	"errors"
+	"fmt"
+	"gopkg.in/ldap.v3"
 )
 
-func LdapConnenct(l *model.Ldap, user string, pass string, isTest bool) (isOk bool, err error) {
+func LdapContent(l *model.Ldap, user string, pass string, isTest bool) (isOk bool, err error) {
 
 	var s string
 
-	ld, err := ldap.Dial("tcp", l.Url)
+	var ld *ldap.Conn
 
 	if l.Ldaps {
-		if err := ld.StartTLS(&tls.Config{InsecureSkipVerify: true}); err != nil {
-			return false, err
-		}
+		ld, err = ldap.DialTLS("tcp", l.Url, &tls.Config{InsecureSkipVerify: true})
+	} else {
+		ld, err = ldap.Dial("tcp", l.Url)
 	}
 
 	if err != nil {
@@ -70,4 +70,3 @@ func LdapConnenct(l *model.Ldap, user string, pass string, isTest bool) (isOk bo
 	}
 	return true, nil
 }
-
